Add 4bit flux kontext loading and `latent_fixed_size`  for 16GB vram training

+++ jobs/process/BaseSDTrainProcess.py
@@ -1589,7 +1589,7 @@ class BaseSDTrainProcess(BaseTrainProcess):
         else:
             text_encoder.requires_grad_(False)
             text_encoder.eval()
-        unet.to(self.device_torch, dtype=dtype)
+        unet.to(self.device_torch)
         unet.requires_grad_(False)
         unet.eval()
         vae = vae.to(torch.device('cpu'), dtype=dtype)
